move constant inference to its own file and data structure
