+++ compiler-core/src/type_.rs
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2018 The Gleam contributors
 
+pub(crate) mod constant;
 pub(crate) mod environment;
 pub mod error;
 pub(crate) mod expression;
